Initial setup

+++ artifacts/api-server/src/routes/activity-logs.ts
@@ -9,7 +9,7 @@ router.get("/activity-logs", async (req, res) => {
   const limit = Math.min(100, Math.max(1, Number(req.query.limit) || 50));
   const offset = (page - 1) * limit;
 
-  const [{ total }] = await db.select({ total: sql<number>`count(*)::int` }).from(activityLogsTable);
+  const [{ total }] = await db.select({ total: sql<number>`count(*)` }).from(activityLogsTable);
   const logs = await db.select().from(activityLogsTable).orderBy(desc(activityLogsTable.createdAt)).limit(limit).offset(offset);
   res.json({ logs, total, page, limit });
 });
